+++ src/windows/editor_object_selection.c
@@ -1694,7 +1694,7 @@ static void window_editor_object_selection_manage_tracks()
 	gS6Info->editor_step = EDITOR_STEP_TRACK_DESIGNS_MANAGER;
 
 	int entry_index = 0;
-	for (; ((int)object_entry_groups[0].chunks[entry_index]) == -1; ++entry_index);
+	for (; ((intptr_t)object_entry_groups[0].chunks[entry_index]) == -1; ++entry_index);
 
 	rct_ride_entry* ride_entry = get_ride_entry(entry_index);
 	uint8* ride_type_array = &ride_entry->ride_type[0];
